hub/requirements.txt
asgiref
Django
djangorestframework
gunicorn
packaging
pika
python-dotenv
sqlparse
psycopg2-binary
gunicorn
pytz
django-debug-toolbar
django-cors-headers
lifxlan